Start implementing TextInputV2
