+++ packages/components/src/TextInput/__stories__/TextInputV2.stories.js
@@ -2,6 +2,7 @@ import React from 'react';
 
 import { Container, Grid, Text } from '../../index';
 import { NumberInput } from './NumberInput';
+import { PresetInput } from './PresetInput';
 import { TextInput } from './TextInput';
 import { UnitInput } from './UnitInput';
 
@@ -11,6 +12,24 @@ export default {
 
 const Example = () => {
 	const [value, setValue] = React.useState('123');
+
+	const presets = [
+		{
+			label: 'Tiny Tiny',
+			key: 'small',
+			value: '9px',
+		},
+		{
+			label: 'Medium Sized',
+			key: 'medium',
+			value: '17px',
+		},
+		{
+			label: 'Largeeeee',
+			key: 'large',
+			value: '28px',
+		},
+	];
 
 	return (
 		<Container width={480}>
@@ -33,8 +52,19 @@ const Example = () => {
 			<Grid>
 				<Text>Unit</Text>
 				<UnitInput
+					cssProp="margin"
 					onBlur={() => console.log('blur')}
 					onChange={setValue}
+					value={value}
+				/>
+			</Grid>
+			<Grid>
+				<Text>Preset</Text>
+				<PresetInput
+					cssProp="margin"
+					onBlur={() => console.log('blur')}
+					onChange={setValue}
+					presets={presets}
 					value={value}
 				/>
 			</Grid>
