+++ packages/components/src/TextInput/index.js
@@ -1,1 +1,2 @@
 export { default as TextInput } from './TextInput';
+export { default as TextInputArrows } from './TextInputArrows';
