+++ packages/utils/src/index.js
@@ -5,17 +5,21 @@ export * from './colors';
 export * from './deepEqual';
 export * from './deepMerge';
 export * from './env';
+export * from './eventHandlers';
 export * from './hooks';
 export * from './is';
+export * from './keyboard';
 export * from './math';
 export * from './memoize';
 export * from './objects';
-export * from './shallowEqual';
 export * from './other';
 export * from './react';
+export * from './selection';
+export * from './shallowEqual';
 export * from './strings';
-export * from './values';
 export * from './unitValues';
+export * from './validation';
+export * from './values';
 export * from './warning';
 
 // Deprecated. use @wp-g2/substate
