+++ packages/utils/src/keyboard.js
@@ -1,0 +1,8 @@
+export function normalizeArrowKey(event) {
+	const { key, keyCode } = event;
+	/* istanbul ignore next (ie) */
+	if (keyCode >= 37 && keyCode <= 40 && key.indexOf('Arrow') !== 0) {
+		return `Arrow${key}`;
+	}
+	return key;
+}
