+++ packages/utils/src/react.js
@@ -5,24 +5,6 @@ import { is } from './is';
 export { default as mergeRefs } from 'react-merge-refs';
 export { default as hoistNonReactStatics } from 'hoist-non-react-statics';
 
-export function mergeEvent(handler, htmlHandler) {
-	return (event) => {
-		if (is.function(htmlHandler)) {
-			htmlHandler(event);
-		}
-		if (is.function(handler)) {
-			handler(event);
-		}
-	};
-}
-
-export function useMergeEventCallback(handler, htmlHandler) {
-	return React.useCallback(mergeEvent(handler, htmlHandler), [
-		handler,
-		htmlHandler,
-	]);
-}
-
 export function getValidChildren(children) {
 	if (is.string(children)) return [children];
